V0.02.03: bake in live-debug fixes from the validated V100 Stage-1 run
Stage-1 hybrid trained end-to-end on V100 (2000 steps, σ held throughout, γ free
to ~0.083). Absorbing the source patches applied via sed during that run so the
next run has them baked in. (#4 write_proj random-init already landed in V0.02.02.)

#1 train.py — γ fp32 master parameter: re-cast model.transformer.gamma to fp32
   AFTER model.to(dtype) and BEFORE the optimizer. model.to casts every param
   regardless of declared dtype; bf16 γ has ULP ~2.4e-4 > AdamW step 1e-4, snaps
   to 1/32 and freezes. This is THE fix that let γ move.
#2 transformer.py — γ declared nn.Parameter(zeros(1, dtype=float32))
   (belt-and-suspenders; train.py recast is load-bearing).
#3 transformer.py inject() — fp32 residual add keeps γ's gradient path fp32
   end-to-end; identity at γ=0 (warm-start preserved exactly).
#5 rwkv_cell.reset() — state zeros follow param dtype (fp32 state vs bf16 params
   crashed `r @ S`). Same latent bug fixed in vru_cell.reset() for parity.

README: record the validated run + the γ-fp32 / Recursion-fp32 / substrate-state
dtype requirements as critical training notes.

Pending: chat.py (inference helper) lands when Dylan sends it; must call
model.recursion.float() after model.to(bf16), same pattern as train.py.

Verified here (no torch): all modules byte-compile; aeon/ transformers-free;
substrate conformance torch-free checks pass.

+++ aeon/transformer.py
@@ -308,7 +308,12 @@ class HybridTransformer(nn.Module):
         # write surface: manifold (H_rec) -> hidden (D), γ-gated from 0 (warm start)
         self.write_proj = nn.Linear(h_rec, self.D, bias=False)
         nn.init.normal_(self.write_proj.weight, std=0.02)
-        self.gamma = nn.Parameter(torch.zeros(1))
+        # γ is an fp32 MASTER parameter. It must stay fp32: a bf16 γ has ULP
+        # 2^-12≈2.4e-4 near 2^-5, above the AdamW step (1e-4), so it snaps to
+        # 1/32 and freezes. Declared fp32 here (belt-and-suspenders); train.py
+        # also re-casts to fp32 AFTER model.to(dtype), which is the load-bearing
+        # fix (model.to casts every param regardless of declared dtype).
+        self.gamma = nn.Parameter(torch.zeros(1, dtype=torch.float32))
 
     # ---- weight init from R1 ---------------------------------------------
     def load_pretrained(self, checkpoint_dir: str) -> dict:
@@ -330,7 +335,10 @@ class HybridTransformer(nn.Module):
         return self.read_proj(hidden)
 
     def inject(self, hidden, signal):
-        return hidden + self.gamma * self.write_proj(signal)
+        # fp32 residual add keeps γ's gradient path fp32 from loss back to the
+        # parameter (protect the gate signal from bf16 quantization). At γ=0 this
+        # is identity: hidden.float() + 0, cast back to hidden's dtype exactly.
+        return (hidden.float() + self.gamma * self.write_proj(signal).float()).to(hidden.dtype)
 
     def logits(self, hidden):
         return self.model.logits(hidden)
